requirements.txt
numpy
torch>=1.0
torchvision
matplotlib
tensorflow
tensorboard==1.14.0
terminaltables
pillow
tqdm
